Refactor brews to build the new section and make it independent from the pill
