+++ app/src/main/java/com/brewthings/app/data/model/Brews.kt
@@ -1,0 +1,6 @@
+package com.brewthings.app.data.model
+
+data class Brews(
+    val batchName: String,
+    val data: List<Brew>,
+)
