+++ app/src/main/java/com/brewthings/app/di/Koin.kt
@@ -1,8 +1,10 @@
 package com.brewthings.app.di
 
 import com.brewthings.app.data.ble.RaptPillScanner
+import com.brewthings.app.data.repository.BrewsRepository
 import com.brewthings.app.data.repository.RaptPillRepository
 import com.brewthings.app.data.storage.RaptPillDatabase
+import com.brewthings.app.ui.screens.brews.BrewsScreenViewModel
 import com.brewthings.app.ui.screens.pill.GraphScreenViewModel
 import com.brewthings.app.ui.screens.scanning.ScanningScreenViewModel
 import org.koin.android.ext.koin.androidContext
@@ -10,10 +12,14 @@ import org.koin.androidx.viewmodel.dsl.viewModel
 import org.koin.dsl.module
 
 val appModule = module {
-    factory { RaptPillScanner() }
     single { RaptPillDatabase.create(context = androidContext()) }
+
+    factory { RaptPillScanner() }
     factory { get<RaptPillDatabase>().raptPillDao() }
     factory { RaptPillRepository(scanner = get(), dao = get()) }
+    factory { BrewsRepository(dao = get()) }
+
     viewModel { ScanningScreenViewModel() }
+    viewModel { BrewsScreenViewModel() }
     viewModel { GraphScreenViewModel() }
 }
