+++ app/src/main/java/com/brewthings/app/ui/components/VerticalSpace.kt
@@ -1,0 +1,12 @@
+package com.brewthings.app.ui.components
+
+import androidx.compose.foundation.layout.Spacer
+import androidx.compose.foundation.layout.size
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.unit.dp
+
+@Composable
+fun VerticalSpace() {
+    Spacer(modifier = Modifier.size(16.dp))
+}
