+++ app/src/main/java/com/brewthings/app/ui/screens/brews/BrewsScreenState.kt
@@ -1,0 +1,5 @@
+package com.brewthings.app.ui.screens.brews
+
+import com.brewthings.app.data.model.Brews
+
+data class BrewsScreenState(val brews: List<Brews> = emptyList())
