+++ app/src/main/java/com/brewthings/app/ui/screens/navigation/legacy/Destination.kt
@@ -3,6 +3,7 @@ package com.brewthings.app.ui.screens.navigation.legacy
 object Destination {
     const val HOME = "home"
     const val PILL_GRAPH = "pill_graph"
+    const val BREW_GRAPH = "brew_graph"
 
     /* Deep links to specific tabs
     const val SCANNING = "$HOME/${Tab.SCANNING}"
