+++ app/src/main/java/com/brewthings/app/ui/screens/navigation/legacy/HomeScreen.kt
@@ -82,7 +82,7 @@ fun HomeScreen(
                 )
             }
             composable(Destination.Tab.BREWS) {
-                BrewsScreen()
+                BrewsScreen(navController = outerNavController)
             }
         }
     }
