+++ app/src/main/java/com/brewthings/app/ui/screens/navigation/legacy/ParameterHolders.kt
@@ -1,12 +1,18 @@
 package com.brewthings.app.ui.screens.navigation.legacy
 
+import com.brewthings.app.data.model.Brew
+
 /**
  * A holder for parameters that are passed between screens.
  * Jetpack Compose navigation sucks and I hate Google.
  */
-object ParameterHolder {
-    object Graph {
+object ParameterHolders {
+    object PillGraph {
         var name: String? = null
         var macAddress: String? = null
+    }
+
+    object BrewGraph {
+        var brew: Brew? = null
     }
 }
