+++ app/src/main/java/com/brewthings/app/util/Utils.kt
@@ -1,5 +1,10 @@
 package com.brewthings.app.util
 
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.mutableStateOf
+import androidx.compose.runtime.remember
+import androidx.compose.runtime.setValue
 import java.util.Locale
 
 fun ByteArray.asHexString(prefix: String = "0x"): String = joinToString(
@@ -23,3 +28,17 @@ inline fun <T, C : List<T>> C.onEachReverse(
 }
 
 fun Float.sumAll(others: Collection<Float>): Float = this + others.sum()
+
+@Composable
+fun <T : Any?> newOrCached(
+    data: T,
+    initialValue: T,
+): T {
+    var previousData: T by remember { mutableStateOf(initialValue) }
+    return if (data != null) {
+        previousData = data
+        data
+    } else {
+        previousData
+    }
+}
